requirements.txt
scapy==2.5.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.9.0
psutil==5.9.5
pyyaml==6.0.1
colorama==0.4.6
tqdm==4.66.1
joblib==1.3.2
